Final bugfixes. Added Readme.

+++ Assignment 1/question3.py
@@ -1,7 +1,7 @@
 # Question 3
 # I think it is pretty self explanatory
 
-m = int(input("Enter number m here: "))
+m = float(input("Enter number m here: "))
 n = 0
 while n**2 <= m:
     n += 1
